requirements.txt
openai
azure-cognitiveservices-speech

streamlit
